move firework damage listener to MiscListener
